[Fix][Core] Close unused pipe file descriptor of child processes of Raylet
Closes: ray-project/ray#52385
