+++ src/ray/util/process.h
@@ -40,6 +40,23 @@ enum { PID_MAX_LIMIT = 1 << 22 };
 
 namespace ray {
 
+#if defined(_WIN32)
+// Don't care what exact type is in windows... Looks like to be an asio specific type.
+template <typename NativeHandleType>
+void SetFdCloseOnExec(const NativeHandleType &handle);
+#else
+
+/// Sets the flag FD_CLOEXEC to a file descriptor.
+/// This means when the process is forked, this fd would be closed in the child process
+/// side. Raylet forks to create core workers and we don't want the Unix Socket FDs to be
+/// inherited by the core workers. Leaking these FDs would have performance implications.
+///
+/// Idempotent. Calling twice == calling once.
+/// Not thread safe.
+/// See https://github.com/ray-project/ray/issues/40813
+void SetFdCloseOnExec(int fd);
+#endif
+
 class EnvironmentVariableLess {
  public:
   bool operator()(char a, char b) const;
